Move external_call into separate module.
